move some state logic to baseState
